Convert db into a db module.
