+++ integration/regtest-node.js
@@ -27,6 +27,7 @@ var index = require('..');
 var BitcoreNode = index.Node;
 var AddressModule = index.modules.AddressModule;
 var BitcoinModule = index.modules.BitcoinModule;
+var DBModule = index.modules.DBModule;
 var testWIF = 'cSdkPxkAjA4HDr5VHgsebAPDEh9Gyub4HK8UJr2DFGGqKKy4K5sG';
 var testKey;
 var client;
@@ -65,6 +66,11 @@ describe('Node Functionality', function() {
         datadir: datadir,
         network: 'regtest',
         modules: [
+          {
+            name: 'db',
+            module: DBModule,
+            dependencies: DBModule.dependencies
+          },
           {
             name: 'bitcoind',
             module: BitcoinModule,
