+++ lib/chain.js
@@ -99,30 +99,25 @@ Chain.prototype.initialize = function() {
   var self = this;
 
   // Does our database already have a tip?
-  self.node.db.getMetadata(function getMetadataCallback(err, metadata) {
+  self.node.modules.db.getMetadata(function getMetadataCallback(err, metadata) {
     if(err) {
       return self.emit('error', err);
     } else if(!metadata || !metadata.tip) {
       self.tip = self.genesis;
       self.tip.__height = 0;
       self.tip.__weight = self.genesisWeight;
-      self.node.db.putBlock(self.genesis, function putBlockCallback(err) {
+      self.node.modules.db.connectBlock(self.genesis, function(err) {
         if(err) {
           return self.emit('error', err);
         }
-        self.node.db._onChainAddBlock(self.genesis, function(err) {
-          if(err) {
-            return self.emit('error', err);
-          }
 
-          self.emit('addblock', self.genesis);
-          self.saveMetadata();
-          self.emit('initialized');
-        });
+        self.emit('addblock', self.genesis);
+        self.saveMetadata();
+        self.emit('initialized');
       });
     } else {
       metadata.tip = metadata.tip;
-      self.node.db.getBlock(metadata.tip, function getBlockCallback(err, tip) {
+      self.node.modules.db.getBlock(metadata.tip, function getBlockCallback(err, tip) {
         if(err) {
           return self.emit('error', err);
         }
@@ -215,7 +210,7 @@ Chain.prototype.getHashes = function getHashes(tipHash, callback) {
         }
       } else {
         // do a db call if we don't have it
-        self.node.db.getPrevHash(hash, function(err, prevHash) {
+        self.node.modules.db.getPrevHash(hash, function(err, prevHash) {
           if(err) {
             return callback(err);
           }
@@ -246,7 +241,7 @@ Chain.prototype.saveMetadata = function saveMetadata(callback) {
 
   self.lastSavedMetadata = new Date();
 
-  self.node.db.putMetadata(metadata, callback);
+  self.node.modules.db.putMetadata(metadata, callback);
 };
 
 module.exports = Chain;
