+++ lib/modules/address.js
@@ -82,10 +82,10 @@ AddressModule.prototype.transactionOutputHandler = function(messages, tx, output
   }
 
   // Find the address for the output
-  var address = script.toAddress(this.node.db.network);
+  var address = script.toAddress(this.node.network);
   if (!address && script.isPublicKeyOut()) {
     var pubkey = script.chunks[0].buf;
-    address = Address.fromPublicKey(new PublicKey(pubkey), this.node.db.network);
+    address = Address.fromPublicKey(new PublicKey(pubkey), this.node.network);
   } else if (!address){
     return;
   }
@@ -162,10 +162,10 @@ AddressModule.prototype.blockHandler = function(block, addOutput, callback) {
         continue;
       }
 
-      var address = script.toAddress(this.node.db.network);
+      var address = script.toAddress(this.node.network);
       if (!address && script.isPublicKeyOut()) {
         var pubkey = script.chunks[0].buf;
-        address = Address.fromPublicKey(new PublicKey(pubkey), this.node.db.network);
+        address = Address.fromPublicKey(new PublicKey(pubkey), this.node.network);
       } else if (!address){
         continue;
       }
@@ -329,7 +329,7 @@ AddressModule.prototype.getOutputs = function(addressStr, queryMempool, callback
   var outputs = [];
   var key = [AddressModule.PREFIXES.OUTPUTS, addressStr].join('-');
 
-  var stream = this.node.db.store.createReadStream({
+  var stream = this.node.modules.db.store.createReadStream({
     start: key,
     end: key + '~'
   });
@@ -443,7 +443,7 @@ AddressModule.prototype.getSpendInfoForOutput = function(txid, outputIndex, call
   var self = this;
 
   var key = [AddressModule.PREFIXES.SPENTS, txid, outputIndex].join('-');
-  this.node.db.store.get(key, function(err, value) {
+  this.node.modules.db.store.get(key, function(err, value) {
     if(err) {
       return callback(err);
     }
@@ -492,12 +492,12 @@ AddressModule.prototype.getAddressHistoryForAddress = function(address, queryMem
       return callback(null, txinfos[txid]);
     }
 
-    self.node.db.getTransactionWithBlockInfo(txid, queryMempool, function(err, transaction) {
+    self.node.modules.db.getTransactionWithBlockInfo(txid, queryMempool, function(err, transaction) {
       if(err) {
         return callback(err);
       }
 
-      transaction.populateInputs(self.node.db, [], function(err) {
+      transaction.populateInputs(self.node.modules.db, [], function(err) {
         if(err) {
           return callback(err);
         }
