+++ lib/modules/bitcoind.js
@@ -1,14 +1,15 @@
 'use strict';
 
 var util = require('util');
-var Module = require('../module');
 var bindings = require('bindings')('bitcoind.node');
 var mkdirp = require('mkdirp');
 var fs = require('fs');
-var index = require('../');
-var log = index.log;
 var bitcore = require('bitcore');
 var $ = bitcore.util.preconditions;
+var index = require('../');
+var log = index.log;
+var Module = require('../module');
+
 
 /**
  * Provides an interface to native bindings to Bitcoin Core
