+++ lib/node.js
@@ -1,10 +1,8 @@
 'use strict';
 
-var fs = require('fs');
 var util = require('util');
 var EventEmitter = require('events').EventEmitter;
 var async = require('async');
-var mkdirp = require('mkdirp');
 var bitcore = require('bitcore');
 var BufferUtil = bitcore.util.buffer;
 var Networks = bitcore.Networks;
@@ -12,7 +10,6 @@ var _ = bitcore.deps._;
 var $ = bitcore.util.preconditions;
 var Block = bitcore.Block;
 var Chain = require('./chain');
-var DB = require('./db');
 var index = require('./');
 var log = index.log;
 var Bus = require('./bus');
@@ -23,7 +20,6 @@ function Node(config) {
     return new Node(config);
   }
 
-  this.db = null;
   this.chain = null;
   this.network = null;
 
@@ -81,7 +77,7 @@ Node.prototype.addModule = function(service) {
 };
 
 Node.prototype.getAllAPIMethods = function() {
-  var methods = this.db.getAPIMethods();
+  var methods = [];
   for(var i in this.modules) {
     var mod = this.modules[i];
     methods = methods.concat(mod.getAPIMethods());
@@ -90,7 +86,7 @@ Node.prototype.getAllAPIMethods = function() {
 };
 
 Node.prototype.getAllPublishEvents = function() {
-  var events = this.db.getPublishEvents();
+  var events = [];
   for (var i in this.modules) {
     var mod = this.modules[i];
     events = events.concat(mod.getPublishEvents());
@@ -100,8 +96,6 @@ Node.prototype.getAllPublishEvents = function() {
 
 Node.prototype._loadConfiguration = function(config) {
   this._loadNetwork(config);
-  this._loadDB(config);
-  this._loadAPI();
   this._loadConsensus(config);
 };
 
@@ -185,7 +179,7 @@ Node.prototype._syncBitcoindRewind = function(block, done) {
           }
 
           // Undo the related indexes for this block
-          self.db._onChainRemoveBlock(tip, function(err) {
+          self.modules.db.disconnectBlock(tip, function(err) {
             if (err) {
               return removeDone(err);
             }
@@ -258,7 +252,7 @@ Node.prototype._syncBitcoind = function() {
             return done(err);
           }
           // Create indexes
-          self.db._onChainAddBlock(block, function(err) {
+          self.modules.db.connectBlock(block, function(err) {
             if (err) {
               return done(err);
             }
@@ -325,39 +319,6 @@ Node.prototype._loadNetwork = function(config) {
   $.checkState(this.network, 'Unrecognized network');
 };
 
-Node.prototype._loadDB = function(config) {
-  var options = _.clone(config.db || {});
-
-  if (config.DB) {
-    // Other modules can inherit from our DB and replace it with their own
-    DB = config.DB;
-  }
-
-  // Store the additional indexes in a new directory
-  // based on the network configuration and the datadir
-  $.checkArgument(config.datadir, 'Please specify "datadir" in configuration options');
-  $.checkState(this.network, 'Network property not defined');
-  var regtest = Networks.get('regtest');
-  if (this.network === Networks.livenet) {
-    options.path = config.datadir + '/bitcore-node.db';
-  } else if (this.network === Networks.testnet) {
-    options.path = config.datadir + '/testnet3/bitcore-node.db';
-  } else if (this.network === regtest) {
-    options.path = config.datadir + '/regtest/bitcore-node.db';
-  } else {
-    throw new Error('Unknown network: ' + this.network);
-  }
-  options.network = this.network;
-
-  if (!fs.existsSync(options.path)) {
-    mkdirp.sync(options.path);
-  }
-
-  options.node = this;
-
-  this.db = new DB(options);
-};
-
 Node.prototype._loadConsensus = function(config) {
   var options;
   if (!config) {
@@ -369,24 +330,9 @@ Node.prototype._loadConsensus = function(config) {
   this.chain = new Chain(options);
 };
 
-Node.prototype._loadAPI = function() {
-  var self = this;
-  var methodData = self.db.getAPIMethods();
-  methodData.forEach(function(data) {
-    var name = data[0];
-    var instance = data[1];
-    var method = data[2];
-
-    self[name] = function() {
-      return method.apply(instance, arguments);
-    };
-  });
-};
-
 Node.prototype._initialize = function() {
   var self = this;
 
-  this._initializeDatabase();
   this._initializeChain();
 
   this.start(function(err) {
@@ -394,18 +340,6 @@ Node.prototype._initialize = function() {
       return self.emit('error', err);
     }
     self.emit('ready');
-  });
-};
-
-Node.prototype._initializeDatabase = function() {
-  var self = this;
-  this.db.on('ready', function() {
-    log.info('Bitcoin Database Ready');
-  });
-
-  this.db.on('error', function(err) {
-    Error.captureStackTrace(err);
-    self.emit('error', err);
   });
 };
 
@@ -433,10 +367,6 @@ Node.prototype._initializeChain = function() {
 
 Node.prototype.getServices = function() {
   var services = [
-    {
-      name: 'db',
-      dependencies: ['bitcoind'],
-    },
     {
       name: 'chain',
       dependencies: ['db']
