+++ lib/scaffold/default-config.js
@@ -13,7 +13,7 @@ function getDefaultConfig() {
       datadir: process.env.BITCORENODE_DIR || path.resolve(process.env.HOME, '.bitcoin'),
       network: process.env.BITCORENODE_NETWORK || 'livenet',
       port: process.env.BITCORENODE_PORT || 3001,
-      modules: ['bitcoind', 'address']
+      modules: ['bitcoind', 'db', 'address']
     }
   };
 }
